ref(ExamService): move convertExamsToDTOs
