+++ backend/src/main/java/com/Uniquest/UniQuest/controllers/ExamController.java
@@ -9,7 +9,6 @@ import com.Uniquest.UniQuest.dto.CommentResponseDTO;
 import com.Uniquest.UniQuest.dto.ExamResponseDTO;
 import com.Uniquest.UniQuest.dto.ExamTextRequestDTO;
 import com.Uniquest.UniQuest.dto.QuestionDTO;
-import com.Uniquest.UniQuest.service.ExamDTOService;
 import com.Uniquest.UniQuest.service.ExamService;
 import com.Uniquest.UniQuest.service.InteractionUserService;
 import com.Uniquest.UniQuest.service.UserService;
@@ -36,7 +35,6 @@ public class ExamController {
 
     private final ExamService examService;
     private final ExamRepository examRepository;
-    private final ExamDTOService examDTOService;
 
     @PostMapping("/upload/image")
     public ResponseEntity<String> uploadImageExam(@RequestParam String title,
@@ -143,7 +141,7 @@ public class ExamController {
     public ResponseEntity<List<ExamResponseDTO>> getAllExams() {
         List<Exam> exams = examService.getAllExams();
 
-        List<ExamResponseDTO> examDTOs = examDTOService.convertExamsToDTOs(exams);
+        List<ExamResponseDTO> examDTOs = examService.convertExamsToDTOs(exams);
 
         return ResponseEntity.ok(examDTOs);
     }
