+++ backend/src/main/java/com/Uniquest/UniQuest/controllers/UserController.java
@@ -5,7 +5,6 @@ import com.Uniquest.UniQuest.domain.user.User;
 import com.Uniquest.UniQuest.dto.*;
 import com.Uniquest.UniQuest.infra.security.TokenService;
 import com.Uniquest.UniQuest.repositories.UserRepository;
-import com.Uniquest.UniQuest.service.ExamDTOService;
 import com.Uniquest.UniQuest.service.ExamService;
 import com.Uniquest.UniQuest.service.InteractionUserService;
 import com.Uniquest.UniQuest.service.UserService;
@@ -35,7 +34,6 @@ public class UserController {
     private final TokenService tokenService;
     private final PasswordResetService passwordResetService;
     private final ExamService examService;
-    private final ExamDTOService examDTOService;
 
     @GetMapping
     public ResponseEntity<?> getUserInfo(HttpServletRequest request) {
@@ -148,7 +146,7 @@ public class UserController {
     public ResponseEntity<List<ExamResponseDTO>> getUserExams(@PathVariable String userId) {
         List<Exam> exams = examService.getExamsByUser(userId); // Busca as provas do usuário
 
-        List<ExamResponseDTO> examDTOs = examDTOService.convertExamsToDTOs(exams);
+        List<ExamResponseDTO> examDTOs = examService.convertExamsToDTOs(exams);
 
         return ResponseEntity.ok(examDTOs);
     }
